+++ src/utils/index.ts
@@ -22,7 +22,7 @@ import {
   wiki
 } from "./commonUtils";
 import { drawCellsValue, drawPath, drawPoint, drawPolygons, drawRouteConnections } from "./debugUtils";
-import { downloadFile, getFileName, uploadFile } from "./fileUtils";
+import { downloadFile, getFileName, isImageIcon, sanitizeSvgIcon, svgToDataUri, uploadFile } from "./fileUtils";
 import { distanceSquared, rollups } from "./functionUtils";
 import { isLand, isWater, SEA_LEVEL } from "./heightUtils";
 import { applyOption, ensureEl, findEl, getComposedPath, getNextId, getPointer } from "./nodeUtils";
@@ -192,6 +192,7 @@ export {
   getVertexPath,
   initializePrompt,
   isCtrlClick,
+  isImageIcon,
   isLand,
   isValidJSON,
   isVowel,
@@ -219,10 +220,12 @@ export {
   SEA_LEVEL,
   safeParseJSON,
   sanitizeId,
+  sanitizeSvgIcon,
   setInlineStyleProperty,
   si,
   speak,
   splitInTwo,
+  svgToDataUri,
   TYPED_ARRAY_MAX,
   throttle,
   toCsvField,